harden: fix shift UB, printf specifiers, and OOM null-derefs
Audit hygiene fixes, no behavioral change to either execution path:
- sim_core.c: REG_TOUCH/REG_ISTOUCHED use 1u<< to avoid signed shift-by-31
  UB (trace-only); avr_regname sprintf uses %u for the unsigned reg index.
- sim_interrupts.c: IRQ-register trace printf uses %u with explicit casts
  for the unsigned enable.reg/enable.bit bitfields.
- sim_avr.c: guard the three fatal allocations (SRAM, trace_data,
  avr_core_allocate) against OOM with a logged abort() instead of a NULL
  dereference.

Verified: libsimavr builds clean, engine self-test passes, tests 33/33,
cppcheck clean on the touched files.

+++ simavr/sim/sim_interrupts.c
@@ -78,8 +78,9 @@ avr_register_vector(
 	if (vector->vector > table->max_vector)
 		table->max_vector = vector->vector;
 	if (vector->trace)
-		printf("IRQ%d registered (enabled %04x:%d)\n",
-			vector->vector, vector->enable.reg, vector->enable.bit);
+		printf("IRQ%d registered (enabled %04x:%u)\n",
+			vector->vector, (unsigned int)vector->enable.reg,
+			(unsigned int)vector->enable.bit);
 
 	if (!vector->enable.reg)
 		AVR_LOG(avr, LOG_WARNING, "IRQ%d No 'enable' bit !\n",
